added delete secure

+++ src/Pages/ManageInventories/ManageInventories.js
@@ -1,10 +1,12 @@
+import { useAuthState } from 'react-firebase-hooks/auth';
 import { useNavigate } from 'react-router-dom';
+import auth from '../../firebase.init';
 import useCars from '../../hooks/LoadCars/useCars';
 import AllCars from '../Home/InventoryItems/AllCars';
 
 
 const ManageInventories = () => {
-
+    const [user] = useAuthState(auth)
     const [cars, setCars] = useCars();
     const navigate = useNavigate();
 
@@ -15,20 +17,22 @@ const ManageInventories = () => {
     // delete item
     const handleDelete = id => {
         console.log(id)
-        const proceed = window.confirm('Are you sure?');
-        if (proceed) {
-            const url = `https://mysterious-bayou-78729.herokuapp.com/cars/${id}`;
-            console.log(url)
-            fetch(url, {
-                method: 'DELETE'
-            })
-                .then(res => res.json())
-                .then(data => {
-                    console.log(data);
-                    const remaining = cars.filter(car => car._id !== id);
-                    setCars(remaining);
-                });
-        };
+        if (user) {
+            const proceed = window.confirm('Are you sure?');
+            if (proceed) {
+                const url = `https://mysterious-bayou-78729.herokuapp.com/cars/${id}`;
+                console.log(url)
+                fetch(url, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+                        const remaining = cars.filter(car => car._id !== id);
+                        setCars(remaining);
+                    });
+            };
+        }
     };
 
     return (
